+++ ShinyApp/app.R
@@ -57,6 +57,20 @@ allocations <- drake::readd(SEAPSUTWorkflow::target_names$CompletedAllocationTab
 # Adds a combined Machine-Eu.product column
 allocations$Machine_Eu.product = paste(allocations$Machine," - ", allocations$Eu.product)
 
+# Identifies the unique machine and eu_products
+
+machine_uniq <- as.data.frame(unique(etas_and_phis$Machine))
+
+machine_uniq <- magrittr::set_colnames(machine_uniq, "Machine")
+
+eu_product_uniq <- as.data.frame(unique(etas_and_phis$Eu.product))
+
+eu_product_uniq <- magrittr::set_colnames(eu_product_uniq, "Eu.product")
+
+machine_eu_product_uniq <- eu_product_uniq <- as.data.frame(unique(allocations$Machine_Eu.product))
+
+machine_eu_product_uniq <- magrittr::set_colnames(machine_eu_product_uniq, "Machine - Eu.product")
+
 
 ################################################################################
 
@@ -131,7 +145,7 @@ ui <- dashboardPage(
                 
                 # 6 - Direct hyperlink to Github repository (this is probably not necessary)
                 menuItem("Github Repository", icon = icon("file-code"),
-                         href = "https://github.com/ZekeMarshall/PFU-Interface/")
+                         href = "https://github.com/EnergyEconomyDecoupling/PFU-Database")
                 )),
   
   dashboardBody(
@@ -150,11 +164,15 @@ ui <- dashboardPage(
                     width = 12,
                     tabPanel(
                       title = "Machines",
-                      DT::dataTableOutput(outputId = "machines")
+                      DT::dataTableOutput(outputId = "Machines")
                     ),
                     tabPanel(
                       title = "Useful work products",
                       DT::dataTableOutput(outputId = "Eu.products")
+                    ),
+                    tabPanel(
+                      title = "Machine-Useful Work Product",
+                      DT::dataTableOutput(outputId = "machine_eu_product")
                     )
                     ))),
       
@@ -322,7 +340,8 @@ ui <- dashboardPage(
                                          value = 1960,
                                          step = 1,
                                          sep = "",
-                                         width = "100%"
+                                         width = "100%",
+                                         animate = TRUE
                              ),
                       
                fluidRow(
@@ -591,6 +610,7 @@ selected_data_sankey <- reactive({
 })
 
 # What data should be included here, at the moment it is simply etas and phis, need to add: allocations, ECC, IEAData?
+# I need to remove this data section and have a seperate data tabPanel for each visualisation!
 selected_data_DT <- reactive({
   validate(
     need(input$Quantity_DT != "", "Please select at least one Quantity"),
@@ -681,13 +701,19 @@ output$data_table <- DT::renderDataTable({
 #   #   dplyr::select(Country, Quantity, Last.stage, Unit, Machine, Eu.product, Year, .values)
 # })
 
-# output$machines <- DT::renderDataTable({
-#   machines
-# })
-# 
-# output$Eu.products <- DT::renderDataTable({
-#   Eu.products
-# })
+output$Machines <- DT::renderDataTable({
+  machine_uniq
+})
+
+output$Eu.products <- DT::renderDataTable({
+  eu_product_uniq
+})
+
+output$machine_eu_product <- DT::renderDataTable({
+  machine_eu_product_uniq
+})
+
+#########################################
 
 output$downloadData <- downloadHandler(